Solving day 12, part 2 with a basic optimizing interpreter!
There's way more code here than necessary, and I didn't actually test it to be bug-free
beyond the requirements of the challenge.

However, I believe it correctly detects blocks that repeat (# of register times) and can
apply an optimization to remove the loop.

There are 3 loops in my input that ["inc r1", "dec r2", "jnz r2 -2"], and I probably should
have just identified that *specific* pattern. But this was more fun.