Day 17 part 2, re-purposing AStar search code
Very slow in the playground, even with all the work done outside the playground page.

Reasoned about it and concluded there wasn't any way to prune the search space, and every
location needs to be visited. Reading reddit afterwords reminded me that A* won't be the
most efficient (memory-wise). Oh well, it works.