Solving day 21 part 2
I'm not proud of the way I'm inverting rotateBasedOnLetter(), but this was the first solution
I found that worked - and it runs fast enough.

Clearly it'd be better to calculate once the inverse mapping based on the length of the
password, and then just lookup the rotation amount based on the destination index. Or if
I found an equation that worked, but it wasn't obvious how to invert:
(2 * idx + (2 or 1)) % length == destination index

It also turns out that it's very dependent on the length of the password. 7 characters long
wouldn't be invertible, because you couldn't figure out how far the password had been rotated.