Solving day 22, part 2 manually
I wanted to see what it looked like, so I wrote the (slow) code to print out the datacenter.
The solution was trivial to figure out manually, so I didn't bother writing any more code
to do it. I had previously inspected the Nodes, and saw there wasn't anything interesting
in the sizes of nodes: 1 was empty, and most nodes fit in other nodes.

I suspect a search algorithm could do it, although it might be hard to figure out a
heuristic that efficiently solves it. You'd want to move the 'empty' node between the
Goal data and the Goal Node, directly adjacent to the goal data (without going through the
goal data), unless it was already there, then you'd want to move the goal data.

Of course, now I'm almost interested in figuring it how to write that heuristic, but
probably not interested enough to do it.