Solving day 23 part 2 with no code Changes
This is slow, but still fast enough... One step would be to turn Add/Subtract back on,
and then recognize a nested loop that's doing multiply.

Right now, I don't think I have the patience for it, which is too bad. I really enjoyed
day 12 when I originally did it.