WIP on Day 3 part 2
Made some progress on the bus, not complete but I think I'm on a good
track.

I believe I can correctly map to the canonical ring/side/position for
a location, and that should get me to being able to calculate the
location number, as well as finding the neighbors of a location.