2020 day thirteen part two
Luckily I figured out this was a Diophantine matching the Chinese Remainder
theorem.

Brute force was clearly out, checking millions of #s destroys the Xcode playground.

So, I started with the bus schedules as a system of linear equations. I don't actually
know how to solve them with a computer - and was off reading Swift + Accelerate
documentation, which gets complicated pretty quickly for someone with just a passing
recollection of linear equations.

Something that became ~fairly obvious is that the fact I'm looking for integer solutions
was important. googling for that led me to https://en.wikipedia.org/wiki/Diophantine_equation
and I was pleased to see "Diophantine problems have fewer equations than unknowns and involve finding integers that solve simultaneously all equations.", since I had N+1 unknowns.

Poking through that page, it seemed like the Chinese remainder theorem matched, and I
chose to use the sieve. I didn't even bother reading the general case solution that
involves matrices, and ended up not needing it.

It took a little twiddling to figure out the correct value for the `remainder` so that I
was getting the correct matching sequence, but thanks to the several input examples
I was able to get the right answer on the first submission.