2020 day sixteen part two
Looks like only one of the fields is definitively identified, and it then needs to loop
~19 times to figure out which fields are subsequently determined. Ex: 0 must be A, 1 might
be A or B, ergo 1 has to be B, etc.

I know there's a faster way to do this, but this was definitely fast enough.

The hotspot is actually the ~60k comparisons of every value against every rule to figure
out which rules could not apply for each ticket position. I don't think there's any way
to avoid that work.

I moved *some* shared work into the TicketDocument constructor, instead of repeating it
for p1 and p2, but I think that was probably negligible and not worth the effort.