2020 day seventeen, part one
I went pretty far into the weeds on this one: modeling a closed 3d grid with arbitrary
start coordinates, letting the x/y/z coordinates of each square stay the same between
iterations.

Unfortunately, p2 is to make the cube 4 dimensional, which is *not* one of the degrees of
freedom in these structs.