2020 day eighteen part one
Didn't get around to this yesterday, and today it took a while to brush the rust off my
parser building memories.

If I'm reading/remembering my textbook correctly, this is an LL(1) recursive descent
parser. It uses left recursion to get operator precedence correct.

Not really looking forward to the part 2 tweak, but maybe it won't be terrible