2020 day twenty part one
I was thinking about re-assembling the whole image for p1, lucky for me while exploring
the edge data I realized the corners are easy to identify, since the outside edges don't
have matches. It could have been possible that they did - like if the tiles that were
supposed to be on the outside matched with some other outside edge, but luckily the input
isn't structured that way. There are exactly enough unmatched edges to be the outside of
the image, and all 4 corners have *two* edges that are unmatched, making them really easy
to find