2020 day twenty two, part two
This is essentially a naive solution, with basic memoization at the recursive-game level.
It's still pretty slow, taking at least several minutes to complete.

There might be some math and/or algorithmic shortcut I could take, but at this point my
time is worth more to me than the computer's.