WIP on 2020 Day 23 part 2
all these cups & moves are computationally intensive. Poking at using a LinkedList, but
running to complications with Swift's built-in collection types, and I'm too stubborn
to ignore them.

I think a "fast lookup" from value to Node, for arbitrary insertion into the LinkedList is
what's currently missing, but since the Index type needs to be comparable (including <),
it's "tricky"