2022, Day 11, part one
Spent a _bunch_ of time playing with Swift's RegexBuilder, and trying to get a clean-ish
parse of the Monkey objects.

Way uglier than parsing approaches I've used in the past, of identifying the core numbers
and just splitting the strings on them.

However, I'm mostly happy with it. The way I'm creating a Monkey from the regex match by
passing the match to an initializer seems bad, since there's tight coupling between the
built regexes and the method signature of the un-genericized match object. Probably should
do the extraction in the same code that does the matching.