2022 day 16 part 1
This took me a bunch of time, because I was looking for / trying to come up with a fancy
algorithm to solve the exponential nature of needing to check every path.

Like... TSP with priority prizes! Or Linear Programming.

I did check reddit's solutions pages, and saw that my general approach (djikstra's for
each valuable Valve (or Floyd-Warshall), and stopping after time was reached) was a good
one. It just came down to how to generate the search paths and when to prune.

I also spent a while figuring out how to collapse the graph into just the valuable vertices,
and then realized that was a waste of code & space, since it doesn't matter after computing
the djikstra path costs, and that wasn't the expensive part.

The swift-algorithms code files were part of the thrown away work, but keeping since they
may be useful in future challenges.