2022 Day 19, part 1
I was initially surprised when my first "brute force" attempt was too slow, where I was
building a robot or waiting for one of the unaffordable ones to become affordable.

Further thought led me to the realization that the answer is never building a million Ore
robots, and I got to the idea of putting an upper bound on Ore / Clay / Obsidian robots.
And now it's fast enough - certainly not fast though.

I wonder what part 2 has in store for me...