2022 Day 19, part 2
p2 is super slow on the test input, but reasonably fast on the actual input.

I'd started it running and left. I was surprised when I came back and it had found the
answer. I spent some time tweaking to get better performance, but not a lot.

I might do some more, I think this is probably swift-induced performance problems, and
getting a better data representation might be worth it. Or at least educational for me.