2022, Day 21, part 2
This was fun to figure out. Operators either return a value, or a function that solves for
the unknown `humn` value. That function is the inverse of whatever the operator normally
does, and for the non-commutative operators, there's two versions.

I _think_ I like enum + switch pattern matching for this, even if the function is huge.

I'm _not_ totally happy with the struct hierarchy, I didn't spend much time thinking about
them. My namespacing might be weirder than it could be.

Eliminated the memoization of results since it didn't make a difference.