2022 Day 24, Part 1
Recognized that the full state space + looping blizzards is ~2.5m, and that'll fit in RAM...
So I precompute the entire state space, using my 3D grid (adapted for x,y + time), and
run AStar search.

Modified AStar protocol to allow Goal to be a different type than regular state, since
the goal is an x,y pair ignoring time.

I thought about doing a SceneKit visualization of the state space and the paths being
checked, but it'd be entirely after the solution, and I might be less interested in that
now.