2021, Day 4, part 2
I apparently like doing full sequence of things when AoC asks for both the first and the
last.

So, here's an AnySequence/AnyIterator that produces all board scores in winning order.

I tried using `sequence(state: next:)` as a "newer" alternative, but I think this is
a little more readable: captured state in the AnySequence block, iterative changes in the
AnyIterator block