2021, Day 5, part 2
Pretty straightforward, but dropped `stride(between:..)` because the first solution I had
was so focused on striding by positive steps that it turned all diagonal lines into ones
with positive slopes 🤦‍♂️

So just stride from 1st point to 2nd point, in whatever direction.

Happily this case was included in the example, making it easy to diagnose.