2021, Day 7, Part 2
I don't think it's guaranteed that `mean` of the numbers minimizes the errors, since the
error formula is the triangle numbers, but reddit told me that it was close enough.

Mean minimizes the square of the error, and that's at least much closer to the triangle
numbers.

Covered my bases a little by checking all locations around the mean, instead of just the
floor + ceiling, although in my case it was the same.