2021, Day 15, part 2
I was getting occasional failures, and I hypothesized it was memory corruption.
I'm still skeptical that I have a correctly-working Swift wrapper around BinaryHeap.
Or maybe it was simply instability around order things are allocated and a bug in my
comparison function.

The Heap's ordering should be based purely on the priority, not the contents of the state.
And it'd be nice if that ordering was stable, which I think I was doing with pointer compare