2023, Day 5, part 2
I was off-by-one on the upper bound for Range vs ClosedRange as I wrote p2, otherwise
this went pretty straightforward.

Using SE-0270's RangeSet mostly for the convenience of a collection of Ranges that're
automatically joined, but I think this would work basically the same if it was just an array
of Range objects

Not entirely happy with the _readability_ of `part2`, I think the functional methods
in this case were easy to write, but not necessarily for maintaining.