2023, Day 10, part two
Updated Grid to add a `transformElements` method, generating a new grid with elements
mapped to something else.

Ended up simply using it to create an "empty" grid with the same size and a default element,
and then in-place mutation, because I wanted to figure out which elements aren't part of
the active maze, but I think it ended up okay.

I haven't done anything with winding numbers or points in polygons that I remember, but
wikipedia's high level overview was enough to get me this solution, and I'm happy with it.